Update min/max limits for vboxFromHistogram
Using -PHP_INT_MAX was my first guess at figuring out the issue with the color being fairly far off when testing a single color image where there is only one histogram value.  

I was going to say using the 0-255 limits of RGB values seems like a better idea, but the truth is VBox operates on histogram buckets, so the range is really 0-31.

Also added a comment about the min/max range being for RGB buckets and renamed $rgb to $rgbBuckets.  This passes phpunit tests for me and the color palette returned was the same for several test images including the single color image.